requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langgraph==0.1.7
openai==1.23.2
snowflake-connector-python==3.15.0
pandas==2.2.2
tiktoken==0.6.0
instructor==0.6.5
keybert==0.8.3
scikit-learn==1.4.2
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
PyYAML
beautifulsoup4
ftfy
emoji
langdetect
pyspellchecker
nltk
regex 